Initial Kube Method Implementation
With this method a user can apply multiple multi-document
yaml files specifying config maps and pods.

This handles cases where podman fails, like when the name
of a pod and one of its containers is the same.

This also handles detecting any pods previously deployed
with the same name as one that it is currently trying to
deploy and tears it down.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>