Refactor kube method and add removal code
This commit refactors the kube method code to tear down
pods created by files that were removed in a change first,
then tearing down pods created by files that have been modified
and replacing those pods with new pods created from the contents
of the modified files.

This commit also refactors the code that gets changes and runs
the engine method based on those changes and now the change is
passed to the engineMethod func.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>