Rework git and application processes (#182)
This commit reworks the way fetchit gets changes
and applies them.

The point of this commit is to simplify the fetchit
process in order to make it more clear and extensible.

This commit reduces the previous process for applying
the initial state and following changes to just getting
the current state, getting the latest state, then applying
a method to the difference of those two states. In the case
of an initial apply, the current state is empty. Otherwise
the current state is marked by a git tag on the commit
corresponding to the current state.

This commit also fixes a bug where the scheduler was only
scheduling the first method in the queue to run immediately
rather than all of them.

This commit also removes all calls to the update method
of the fetchit config, since fetchit is using pointers to
manipulate target objects instead of making copies of them.

This commit also adds trailing slashes to all target paths
specifying direcotries, in the configs in the examples
directory.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>