Create CommonMethod struct, extend method interface
This commit creates the CommonMethod struct, which
contains common fields from all the methods. This
struct is embedded in all of the structs that
implement the method interface, this struct also
implements part of the method interface.

This commit extends the method interface to include
the GetTargetPath, SetTarget and SetInitialRun functions
that are implemented by the CommonMethod.

This commit changes the runChangesConcurrent function
back to not being implemented by each individual method.

This commit changes the FetchitConfig type to expect
pointers to arrays for the methods in the type instead
of arrays.

This commit creates the zeroToCurrent function to
be run on the the initial run of every method.

Signed-off-by: Joseph Sawaya <jsawaya@redhat.com>