Add support for reading secret password and PAT
It's considered a fairly universally bad thing to commit credentials
into git.  Implement a more secure solution whereby FetchIt loads
password or PAT from an environment secret (i.e. a special env. var.).
They may be specified either in YAML, or in the environment via
`$FETCHIT_SECRET_PASSWORD` or `$FETCHIT_SECRET_PAT`.  In all cases,
the value indicates the name of an environment-secret to retrieve.  The
environment-secret contains the actual password or PAT to be used.

Signed-off-by: Chris Evich <cevich@redhat.com>